requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
Django==4.2.5
django-cors-headers==4.3.0
django-cors-middleware==1.5.0
djangorestframework==3.14.0
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0


@emotion/react@11.11.1
@emotion/styled@11.11.0
@mui/icons-material@5.14.12
@mui/material@5.14.12
@testing-library/jest-dom@5.17.0
@testing-library/react@13.4.0
@testing-library/user-event@13.5.0
js-cookie@3.0.5
react-dom@18.2.0
react-material-ui-carousel@3.4.2
react-router-dom@6.16.0
react-scripts@5.0.1
react@18.2.0
web-vitals@2.1.4
